refactor [DEV-27]: Rename folder with interfaces
