+++ libs/angular/ui/src/modules/media/guards/media-max.guard.ts
@@ -4,7 +4,7 @@ import { CanActivateFn } from '@angular/router';
 
 import { MediaService } from '../services';
 
-import { MediaBreakpoint } from '../models';
+import { MediaBreakpoint } from '../interfaces';
 
 /**
  * Guard to activate based on the maximum media breakpoint.
