+++ libs/angular/ui/src/modules/media/interfaces/media.interface.ts
