+++ libs/angular/ui/src/modules/media/providers/media-config.provider.ts
@@ -2,7 +2,7 @@ import { Provider } from '@angular/core';
 
 import { MediaService } from '../services';
 
-import { MediaConfigData } from '../models';
+import { MediaConfigData } from '../interfaces';
 
 import { MEDIA_CONFIG } from '../tokens/media.token';
